Avoid Dedicated Server Crash due to calling SERVER side on client-onl… (#33)
* Avoid Dedicated Server Crash due to calling SERVER side on client-only method.

This crash affects my dedicated server, running SpongeForge on 1.12.2. Crashlog is here: https://pastebin.com/atkWTke2 
Probably don't merge this, the "fix" is a hack, but it does get the server running, and I can connect a client to it and play without any negative effects.

* Move sideonly to methods 

Suggested by chocohead: this will not cause class not found exceptions.